Lesson 7- 19/03/25
Today is officially the 7th lesson as I was sick on Monday which was our prac day. I made a massive change in my program as I realised that I coded wrong for the api I was using so I decided to change my api source. I have finished with a read me as I cannot find the error in my code yet.