Lesson 8 (pt2)
Part 2-
I fixed the code for wind and humidity by taking out what was unnecessary and now the system can provide users with temperature, condition, humidity, and wind. It can also detect an error like a spelling mistake and it provides a suggestion to check the input. I also took a screenshot of my README and requirements file to put it into  the Installation under main.py for anyone to be able to observe everything in one file. I have again updated my gantt chart but haven't put it in yet.